requirements.txt
autopep8==1.6.0
colorama==0.4.4
mypy==0.910
mypy-extensions==0.4.3
pycodestyle==2.8.0
toml==0.10.2
typing-extensions==4.0.0
vunit-hdl==4.6.0
